+++ 26.2-rc-2/common/src/main/java/com/iamkaf/bonded/client/HUD.java
@@ -1,6 +1,7 @@
 package com.iamkaf.bonded.client;
 
 import com.iamkaf.bonded.Bonded;
+import com.iamkaf.bonded.BondedClient;
 import com.iamkaf.bonded.block.RepairBenchBlock;
 import com.iamkaf.bonded.block.ToolBenchBlock;
 import com.iamkaf.bonded.leveling.levelers.GearTypeLeveler;
@@ -36,7 +37,6 @@ import java.util.Optional;
 public class HUD {
     public static final int WHITE = 0xffffffff;
     public static final int OUTLINE_COLOR = 0x00000000;
-    public static boolean enabled = true;
     private static Minecraft mc;
 
     /**
@@ -90,7 +90,11 @@ public class HUD {
      */
     private static boolean shouldRender() {
         return !(mc.getDebugOverlay()
-                .showDebugScreen() || mc.gui.hud.isHidden() || !enabled || mc.level == null || mc.player == null);
+                .showDebugScreen()
+                || mc.gui.hud.isHidden()
+                || !BondedClient.CONFIG.benchHudEnabled.get()
+                || mc.level == null
+                || mc.player == null);
     }
 
     /**